// src/interfaces/resource-configuration.interface.ts
export default interface ResourceConfiguration {
    showMatrix: boolean;
    showPages: boolean;
    showAllElements: boolean;
    showAllVotes: boolean;
    showOwnVotes: boolean;
    showComments: boolean;
    canAddElements: boolean;
    canAddAnonymousElements: boolean;
    canRemoveAnyElement: boolean;
    canRemoveOwnElement: boolean;
    canRemoveAnonymousElement: boolean;
    canEditAnyElement: boolean;
    canEditOwnElement: boolean;
    canEditAnonymousElement: boolean;
    canDragAnyElement: boolean;
    canVote: boolean;
    canComment: boolean;
  }